// app/dictionaries/paymentMethods/paymentMethods.d.ts
export type PaymentMethodsLocale = {
  label: {
    pageTitle: string;
    pageDescription: string;
    addCardTitle: string;
    cardNumber: string;
    expiryDate: string;
    securityCode: string;
    addCardButton: string;
    addingCard: string;
    registeredPaymentMethods: string;
    addNewCard: string;
    cancel: string;
    noPaymentMethods: string;
    defaultLabel: string;
    expiryLabel: string;
    nameLabel: string;
    setDefault: string;
    delete: string;
    cannotDelete: string;
    minimumPaymentMethodRequired: string;
    noticeTitle: string;
    notice1: string;
    notice2: string;
    notice3: string;
    deleteConfirm: string;
    statusOpen: string;
    statusPaid: string;
    statusVoid: string;
    statusDraft: string;
    statusUncollectible: string;
    invoicesTitle: string;
    invoicesTableLabel: string;
    invoiceId: string;
    issueDate: string;
    dueDate: string;
    totalAmount: string;
    status: string;
    action: string;
    download: string;
    pageTitle: string;
    pageDescription: string;
    invoicesTabTitle: string;
    paymentMethodsTabTitle: string;
    tabsAriaLabel: string;
  };
  alert: {
    setupIntentFailed: string;
    cardInfoMissing: string;
    unknownError: string;
    defaultSettingFailed: string;
    errorOccurred: string;
    deleteFailed: string;
    cardAlreadyExists: string;
    customerInfoNotFound: string;
    paymentMethodCreationFailed: string;
    setupIntentNotCompleted: string;
    clientSecretNotFound: string;
    allCardInfoRequired: string;
  };
  message: {
    noInvoices: string;
  };
};